project/requirements.txt
django
djangorestframework
django-crispy-forms
requests
django-tables2
jsonpickle
gunicorn
django-environ
pytest-pythonpath
pytest-cov
mixer
pytest
django-debug-toolbar
Pillow
psycopg2-binary==2.9.3
django-cors-headers
djangorestframework-simplejwt